requirements.txt
selenium
webdriver-manager
requests
pywhatkit
python-dotenv
apscheduler